+++ src/components/newses.tsx
@@ -4,7 +4,9 @@ import { AllBlogsQuery } from "../../types/graphql-types"
 import styles from "./newses.module.css"
 
 interface IProps {
+  /** ニュース一覧 */
   data: AllBlogsQuery["newses"]
+  /**  */
   className: string
 }
